Create iterative/single element standard (non-vector) scanners
