+++ Squalr.Engine.Scanning/Scanners/Comparers/Vectorized/SnapshotRegionVectorArrayOfBytesScanner.cs
@@ -100,7 +100,7 @@
             }
 
             // Note that array of bytes must increment by 1 per iteration, unlike data type scans which can increment by vector size
-            for (; this.VectorReadOffset <= this.Region.ElementCount - ByteArraySize; this.VectorReadOffset++)
+            for (; this.VectorReadOffset <= this.Region.Range - ByteArraySize; this.VectorReadOffset++)
             {
                 Vector<Byte> scanResults = this.VectorCompare();
