+++ Squalr.Engine.Scanning/Scanners/ManualScanner.cs
@@ -52,8 +52,8 @@
                             Int32 processedPages = 0;
                             ConcurrentScanBag resultRegions = new ConcurrentScanBag();
 
-                            // ParallelOptions options = ParallelSettings.ParallelSettingsFastest.Clone();
-                            ParallelOptions options = ParallelSettings.ParallelSettingsNone.Clone();
+                            ParallelOptions options = ParallelSettings.ParallelSettingsFastest.Clone();
+                            // ParallelOptions options = ParallelSettings.ParallelSettingsNone.Clone();
                             options.CancellationToken = cancellationToken;
 
                             Parallel.ForEach(
